added logic to create top ten
