+++ Recommendations/Base.lproj/Main.storyboard
@@ -3,7 +3,7 @@
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16086"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -11,21 +11,36 @@
         <!--Recommendations View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="RecommendationsViewController" id="BYZ-38-t0r" customClass="RecommendationsViewController" customModule="Recommendations" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="RecommendationsViewController" id="BYZ-38-t0r" customClass="RecommendationsViewController" customModule="Recommended" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="a68-e5-USx">
-                                <rect key="frame" x="0.0" y="44" width="414" height="818"/>
+                                <rect key="frame" x="0.0" y="88" width="414" height="774"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
                             </tableView>
+                            <navigationBar contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Jyl-4E-Q1h">
+                                <rect key="frame" x="0.0" y="44" width="414" height="44"/>
+                                <items>
+                                    <navigationItem title="Serial Box Titles" id="W3S-LV-lIX">
+                                        <barButtonItem key="rightBarButtonItem" title="Top Ten" id="NJe-7R-jsw">
+                                            <connections>
+                                                <action selector="topTenButtonTapped:" destination="BYZ-38-t0r" id="7OT-WI-h5S"/>
+                                            </connections>
+                                        </barButtonItem>
+                                    </navigationItem>
+                                </items>
+                            </navigationBar>
                         </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
                         <constraints>
-                            <constraint firstItem="a68-e5-USx" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" id="2wg-au-R7E"/>
+                            <constraint firstItem="a68-e5-USx" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="44" id="2wg-au-R7E"/>
                             <constraint firstItem="a68-e5-USx" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" id="6kE-XR-Cnp"/>
+                            <constraint firstItem="Jyl-4E-Q1h" firstAttribute="trailing" secondItem="a68-e5-USx" secondAttribute="trailing" id="ZyP-8e-4OX"/>
                             <constraint firstItem="a68-e5-USx" firstAttribute="bottom" secondItem="6Tk-OE-BBY" secondAttribute="bottom" id="a6z-iY-055"/>
+                            <constraint firstItem="Jyl-4E-Q1h" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" id="awZ-mg-cDd"/>
+                            <constraint firstItem="Jyl-4E-Q1h" firstAttribute="leading" secondItem="a68-e5-USx" secondAttribute="leading" id="cuz-EG-5D2"/>
                             <constraint firstItem="a68-e5-USx" firstAttribute="trailing" secondItem="6Tk-OE-BBY" secondAttribute="trailing" id="hNd-fs-pbh"/>
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
